move merge into seperate file
